started up navigation
