+++ coupon-companion/App.tsx
@@ -3,22 +3,14 @@ import React from 'react';
 import { StyleSheet, Text, View } from 'react-native';
 import { Card } from './components/Card';
 import { CouponCard } from './components/CouponCard';
+import { StackNavigationProp } from '@react-navigation/stack';
+
+export type AppNavigationProp = StackNavigationProp<{ Home: undefined, Login: undefined }>;
 
 export default function App() {
   return (
-    <View style={styles.container}>
-      <Text>Open up App.tsx to start working on your app!</Text>
-      <CouponCard company='Radar.io' details='api calls' discount='20%' textColor='#00e'/>
-      <StatusBar style="auto" />
-    </View>
+
   );
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
+
